+++ backend/core-api/src/modules/logs/graphql/schema/index.ts
@@ -139,6 +139,7 @@ export const queries = `
     logsMainList(${commonQueryParams}):MainLogsList
     logsGetContentTypes: [LogContentType!]!
     logDetail(_id:String!):Log
+    logsRevertPreview(processId: String!): LogRevertResult
 `;
 
 export const mutations = `