+++ frontend/plugins/frontline_ui/src/widgets/automations/modules/facebook/components/bots/components/FacebookPageInfo.tsx
@@ -1,10 +1,17 @@
 import { IconCircleCheck, IconInfoTriangle } from '@tabler/icons-react';
-import { Label } from 'erxes-ui';
+import { Label, Skeleton } from 'erxes-ui';
 import { memo } from 'react';
 import { useFacebookBotPages } from '~/widgets/automations/modules/facebook/components/bots/hooks/useFacebookBotPages';
 
+type FacebookPageInfoProps = {
+  accountId: string;
+  pageId: string;
+};
+
 export const FacebookPageInfo = memo(
-  ({ accountId, pageId }: { accountId: string; pageId: string }) => {
+  ({ accountId, pageId }: FacebookPageInfoProps) => {
+    const { page, loading } = useFacebookBotPages(accountId, pageId);
+
     if (!accountId || !pageId) {
       return (
         <Label className="text-lg flex gap-2 items-center">
@@ -13,7 +20,14 @@ export const FacebookPageInfo = memo(
       );
     }
 
-    const { page } = useFacebookBotPages(accountId, pageId);
+    if (loading) {
+      return (
+        <Label className="text-lg flex gap-2 items-center">
+          <Skeleton className="size-3 rounded-full" />
+          <Skeleton className="h-5 w-36" />
+        </Label>
+      );
+    }
 
     if (!page) {
       return (
@@ -30,7 +44,7 @@ export const FacebookPageInfo = memo(
       </Label>
     );
   },
-  (prevProps: any, nextProps: any) =>
+  (prevProps: FacebookPageInfoProps, nextProps: FacebookPageInfoProps) =>
     prevProps?.accountId === nextProps?.accountId &&
     prevProps?.pageId === nextProps?.pageId,
 );